% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam2.2.R
\docType{data}
\name{DataExam2.2}
\alias{DataExam2.2}
\title{Data for Example 2.2 from Experimental Design and Analysis for Tree Improvement}
\format{
A \code{data.frame} with 16 rows and 2 variables.
\describe{
       \item{\code{Seedlot}}{Two Seedlots Seed Orchad (SO) and routin plantation (P)}
       \item{\code{dbh}}{Diameter at breast height}
       }
}
\usage{
data(DataExam2.2)
}
\description{
Exam2.2 is used to  compare two seed lots by using ANOVA under RCB Design.
}
\examples{
data(DataExam2.2)

}
\references{
\enumerate{
         \item E.R. Williams, C.E. Harwood and A.C. Matheson (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{Exam2.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
