% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam3.1.R
\docType{data}
\name{DataExam3.1}
\alias{DataExam3.1}
\title{Data for Example 3.1 from Experimental Design & Analysis for Tree Improvement}
\format{A \code{data.frame} with 80 rows and 6 variables.
\describe{
       \item{\code{Repl}}{Replication number of different Seedlots}
       \item{\code{PlotNo}}{Plot No of differnt Trees}
       \item{\code{SeedLot}}{Seed Lot number}
       \item{\code{TreeNo}}{Tree number of Seedlots}
       \item{\code{Ht}}{Height in meter}
       \item{\code{Dgl}}{Diameter at ground level}
       }}
\usage{
data(DataExam3.1)
}
\description{
Exam3.1 is part of data from Australian Centre for Agricultural Research (ACIAR) 
             in Queensland, Australia (Experiment 309).
}
\examples{
data(DataExam3.1)

}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{Exam3.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
