% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam4.3.R
\docType{data}
\name{DataExam4.3}
\alias{DataExam4.3}
\title{Data for Example 4.3 from Experimental Design and Analysis for Tree Improvement}
\format{
A \code{data.frame} with 72 rows and 8 variables.
\describe{
       \item{\code{rep}}{Replication number of Treatment}
       \item{\code{row}}{Row number of different Seedlots}
       \item{\code{column}}{Column number of differnt Trees}
       \item{\code{seedlot}}{Seed lot number}
       \item{\code{treat}}{Treatment types}
       \item{\code{count}}{Number of germinated seeds out of 25}
       \item{\code{percent}}{Germination Percentage}
       \item{\code{contcomp}}{Control or Trated Plot}
       }
}
\usage{
data(DataExam4.3)
}
\description{
Exam4.3 presents the germination count data for 4 Pre-Treatments and 6 Seedlots.
}
\examples{
data(DataExam4.3)

}
\references{
\enumerate{
         \item E.R. Williams, C.E. Harwood and A.C. Matheson (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{Exam4.3}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
