% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam5.1.R
\docType{data}
\name{DataExam5.1}
\alias{DataExam5.1}
\title{Data for Example 5.1 from Experimental Design and Analysis for Tree Improvement}
\format{
A \code{data.frame} with 108 rows and 4 variables.
\describe{
       \item{\code{Site}}{Sites for the experiment}
       \item{\code{SeedLot}}{Seed lot number}
       \item{\code{Ht}}{Height of the plants}
       \item{\code{SiteMean}}{Mean Height of Each Site}
         }
}
\usage{
data(DataExam5.1)
}
\description{
Exam5.1 presents the height of 27 seedlots from 4 sites.
}
\examples{
data(DataExam5.1)

}
\references{
\enumerate{
         \item E.R. Williams, C.E. Harwood and A.C. Matheson (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{Exam5.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
