% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam5.2.R
\docType{data}
\name{DataExam5.2}
\alias{DataExam5.2}
\title{Data for Example 5.2 from Experimental Design and Analysis for Tree Improvement}
\format{
A \code{data.frame} with 108 rows and 4 variables.
\describe{
       \item{\code{site}}{Sites for the experiment}
       \item{\code{seedlot}}{Seed lot number}
       \item{\code{ht}}{Height of the plants}
       \item{\code{sitemean}}{Mean Height of Each Site}
         }
}
\usage{
data(DataExam5.2)
}
\description{
Exam5.2 presents the height of 37 seedlots from 6 sites.
}
\examples{
data(DataExam5.2)
}
\references{
\enumerate{
         \item E.R. Williams, C.E. Harwood and A.C. Matheson (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{Exam5.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
