% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam6.2.R
\docType{data}
\name{DataExam6.2}
\alias{DataExam6.2}
\title{Data for Example 6.2 from Experimental Design & Analysis for Tree Improvement}
\format{
A \code{data.frame} with 192 rows and 7 variables.
\describe{
       \item{\code{Replication}}{Replication number of different Families}
       \item{\code{Plot.number}}{Plot number of differnt Trees}
       \item{\code{Family}}{Family Numuber}
       \item{\code{Province}}{Province of family}
       \item{\code{Dbh.mean}}{Average Diameter at breast height of trees within plot}
       \item{\code{Dbh.variance}}{Variance of Diameter at breast height of trees within plot}
       \item{\code{Dbh.count}}{Number of trees within plot}
         }
}
\usage{
data(DataExam6.2)
}
\description{
Exam 6.2 Dbh mean, Dbh varince and number of trees per plot from 3 provinces("PNG","Sabah","Queensland") with 4 replicationsof 48 families.
}
\examples{
data(DataExam6.2)

}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
