% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam3.1.R
\name{Exam3.1}
\alias{Exam3.1}
\title{Data for Example 3.1 from Experimental Design & Analysis for Tree Improvement}
\description{
Exam3.1 is part of data from Australian Centre for Agricultural Research (ACIAR)
             in Queensland, Australia (Experiment 309).
}
\examples{
data(DataExam3.1)
library(tidyverse)

# Pg. 28

fmtab3.3     <-
  lm(
     formula     = Ht ~ Repl*SeedLot
   , data        = DataExam3.1
 # , subset
 # , weights
 # , na.action
   , method      = "qr"
   , model       = TRUE
   , x           = FALSE
   , y           = FALSE
   , qr          = TRUE
   , singular.ok = TRUE
   , contrasts   = NULL
    )

 # Pg. 33

 anova(fmtab3.3)

fmtab3.3ANOVA1 <-
  anova(fmtab3.3) \%>\%
  mutate("F value" = c(anova(fmtab3.3)[1:2, 3]/anova(fmtab3.3)[3, 3], anova(fmtab3.3)[3, 3], NA))

fmtab3.3ANOVA1 \%>\%
  mutate("Pr(>F)"  = pf(q = fmtab3.3ANOVA1[ ,4], df1 = fmtab3.3ANOVA1[ ,1], df2 = fmtab3.3ANOVA1[4,1], lower.tail = FALSE))

library(supernova)

 supernova(fmtab3.3)

 library(emmeans)

 # Pg. 33

 emmeans(
     object     = fmtab3.3
   , specs      = ~ SeedLot
   , by         = NULL
   , fac.reduce = function(coefs) apply(coefs, 2, mean)
   , contr      =
   , options    = get_emm_option("emmeans")
   , weights    =
   , offset     =
   , tran       =
   )
 # Pg. 34
 library(ggplot2)
 ggplot(mapping = aes(x = fitted.values(fmtab3.3), y = residuals(fmtab3.3)))+
 geom_point(size = 2) +
 labs(
    x = "Fitted Values"
  , y = "Residual"
   ) +
 theme_classic()

}
\references{
\enumerate{
         \item Williams, E. R., Matheson, A. C. and Harwood, C. E. (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{DataExam3.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
