% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam2.1.R
\docType{data}
\name{DataExam2.1}
\alias{DataExam2.1}
\title{Data for Example 2.1 from Experimental Design & Analysis for Tree Improvement}
\format{
A \code{data.frame} with 16 rows and 2 variables.
\describe{
       \item{\code{Seedlot}}{Two Seedlots Seed Orchad (SO) and routin plantation (P)}
       \item{\code{dbh}}{Diameter at breast height}
       }
}
\usage{
data(DataExam2.1)
}
\description{
Exam2.1 is used to  compare two seed lots by using single factor ANOVA.
}
\examples{
data(DataExam2.1)

}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{Exam2.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
