% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam4.3.R
\docType{data}
\name{DataExam4.3}
\alias{DataExam4.3}
\title{Data for Example 4.3 from Experimental Design & Analysis for Tree Improvement}
\format{A \code{data.frame} with 72 rows and 8 variables.
\describe{
       \item{\code{Row}}{Row number of different Seedlots}
       \item{\code{Column}}{Column number of differnt Trees}
       \item{\code{Replication}}{Replication number of Treatment}
       \item{\code{Contcomp}}{Control or Trated Plot}
       \item{\code{Pretreatment}}{Treatment types}
       \item{\code{SeedLot}}{Seed lot number}
       \item{\code{GerminationCount}}{Number of germinated seeds out of 25}
       \item{\code{Percent}}{Germination Percentage}
       }}
\usage{
data(DataExam4.3)
}
\description{
Exam4.3 presents the germination count data for 4 Pre-Treatments and 6 Seedlots.
}
\examples{
data(DataExam4.3)

}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{Exam4.3}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
