% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam4.4.R
\docType{data}
\name{DataExam4.4}
\alias{DataExam4.4}
\title{Data for Example 4.4 from Experimental Design & Analysis for Tree Improvement}
\format{A \code{data.frame} with 32 rows and 5 variables.
\describe{
       \item{\code{Rep}}{Replication number}
       \item{\code{Irrig}}{Irrigation type}
       \item{\code{Ferti}}{Fertilizer type}
       \item{\code{SeedDLot}}{Seed Lot number}
       \item{\code{Height}}{Height of the plants}
         }}
\usage{
data(DataExam4.4)
}
\description{
Exam4.4 presents the height means for 4 seedlots under factorial arrangement for two 
             levels of Fertilizer and two levels of Irrigation.
}
\examples{
data(DataExam4.4)

}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{Exam4.4}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
