% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam5.2.R
\docType{data}
\name{DataExam5.2}
\alias{DataExam5.2}
\title{Data for Example 5.2 from Experimental Design & Analysis for Tree Improvement}
\format{A \code{data.frame} with 108 rows and 4 variables.
\describe{
       \item{\code{Site}}{Sites for the experiment}
       \item{\code{SeedLot}}{Seed lot number}
       \item{\code{Ht}}{Height of the plants}
       \item{\code{SiteMean}}{Mean Height of Each Site}
         }}
\usage{
data(DataExam5.2)
}
\description{
Exam5.2 presents the height of 37 seedlots from 6 sites.
}
\examples{
data(Exam5.2)
}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2002).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{Exam5.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
