% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam8.1.R
\docType{data}
\name{DataExam8.1}
\alias{DataExam8.1}
\title{Data for Example 8.1 from Experimental Design and Analysis for Tree Improvement}
\format{
A \code{data.frame} with 236 rows and 8 variables.
\describe{
       \item{\code{repl}}{There are 4 replication for the design}
       \item{\code{row}}{Experiment is conducted under 6 rows}\
       \item{\code{col}}{Experiment is conducted under 4 columns}
       \item{\code{inoc}}{Seedling were inoculated for 2 different time periods half for one week and half for seven weeks}
       \item{\code{prov}}{provenance}
       \item{\code{Country}}{Data for different seedlots was collected from 18 countries}
       \item{\code{Dbh}}{Diameter at breast height}
       \item{\code{Country.1}}{Recoded Country lables}
       }
}
\usage{
data(DataExam8.1)
}
\description{
Exam8.1 presents the Diameter at breast height (Dbh) of 60 SeedLots
             under layout of row column design with 6 rows and 10 columns in 18
             countries and 59 provinces of 18 selected countries.
}
\examples{
data(DataExam8.1)
}
\references{
\enumerate{
         \item E.R. Williams, C.E. Harwood and A.C. Matheson (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{Exam8.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
