% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExam8.2.R
\docType{data}
\name{DataExam8.2}
\alias{DataExam8.2}
\title{Data for Example 8.2 from Experimental Design & Analysis for Tree Improvement}
\format{
A \code{data.frame} with 236 rows and 8 variables.
\describe{
       \item{\code{Replication}}{There are 4 replication for the design}
       \item{\code{Row}}{Experiment is conducted under 6 rows}\
       \item{\code{Column}}{Experiment is conducted under 4 columns}
       \item{\code{Inoc}}{Seedling were inoculated for 2 different time periods half for one week and half for seven weeks}
       \item{\code{Province}}{Experiment was performed in 59 provinces of selected countries}
       \item{\code{Country}}{Data for different seedlots was collected from 18 countries}
       \item{\code{Dbh}}{Diameter at breast height}
       \item{\code{Country.1}}{Recoded Country lables}
       }
}
\usage{
data(DataExam8.2)
}
\description{
Exam8.2 presents the Diameter at breast height (Dbh) of 60 SeedLots
             under layout of row column design with 6 rows and 10 columns in 18
             countries and 59 provinces of 18 selected countries.
}
\examples{
data(DataExam8.2)
}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2023).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{Exam8.2}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
\keyword{datasets}
