% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam4.3.1.R
\name{Exam4.3.1}
\alias{Exam4.3.1}
\title{Example 4.3.1 from Experimental Design & Analysis for Tree Improvement}
\description{
Exam4.3.1 presents the germination count data for 4 Pre-Treatments and 6 Seedlots.
}
\examples{

data(DataExam4.3.1)
library(tidyverse)
library(ggplot2)

# Pg. 51

fm4.4    <-
  aov(
      formula     = Percent ~ Replication + Pretreatment*SeedLot
    , data        = DataExam4.3.1 \%>\%
                    filter(Pretreatment != "Control")
  # , subset
  # , weights
  # , na.action
    , method      = "qr"
    , model       = TRUE
    , x           = FALSE
    , y           = FALSE
    , qr          = TRUE
    , singular.ok = TRUE
    , contrasts   = NULL
   )


 # Pg. 54
 anova(fm4.4)

 library(supernova)

 supernova(fm4.4)

  library(emmeans)


  emmeans(
      object     = fm4.4
    , specs      = ~ Pretreatment
    , by         = NULL
    , fac.reduce = function(coefs) apply(coefs, 2, mean)
    , contr      =
    , options    = get_emm_option("emmeans")
    , weights    =
    , offset     =
    , tran       =
    )

 emmip(
     object        = fm4.4
   , formula       = ~ Pretreatment
   , type          = c("link", "response", "predict.type")[1]
   , CIs           = c(TRUE, FALSE)[1]
   , PIs           = c(TRUE, FALSE)[2]
   , style         =
   , engine        = get_emm_option("graphics.engine")
   , plotit        = TRUE
   , nesting.order = FALSE
   ) +
   theme_classic()

  emmeans(
        object     = fm4.4
      , specs      = ~ SeedLot
      , by         = NULL
      , fac.reduce = function(coefs) apply(coefs, 2, mean)
      , contr      =
      , options    = get_emm_option("emmeans")
      , weights    =
      , offset     =
      , tran       =
      )

 emmip(
     object        = fm4.4
   , formula       = ~ SeedLot
   , type          = c("link", "response", "predict.type")[1]
   , CIs           = c(TRUE, FALSE)[1]
   , PIs           = c(TRUE, FALSE)[2]
   , style         =
   , engine        = get_emm_option("graphics.engine")
   , plotit        = TRUE
   , nesting.order = FALSE
   ) +
   theme_classic()

  emmeans(
        object     = fm4.4
      , specs      = ~ Pretreatment * SeedLot
      , by         = NULL
      , fac.reduce = function(coefs) apply(coefs, 2, mean)
      , contr      =
      , options    = get_emm_option("emmeans")
      , weights    =
      , offset     =
      , tran       =
      )

 emmip(
     object        = fm4.4
   , formula       = SeedLot ~ Pretreatment
   , type          = c("link", "response", "predict.type")[1]
   , CIs           = c(TRUE, FALSE)[1]
   , PIs           = c(TRUE, FALSE)[2]
   , style         =
   , engine        = get_emm_option("graphics.engine")
   , plotit        = TRUE
   , nesting.order = FALSE
   ) +
   theme_classic()


 emmip(
     object        = fm4.4
   , formula       = Pretreatment ~ SeedLot
   , type          = c("link", "response", "predict.type")[1]
   , CIs           = c(TRUE, FALSE)[1]
   , PIs           = c(TRUE, FALSE)[2]
   , style         =
   , engine        = get_emm_option("graphics.engine")
   , plotit        = TRUE
   , nesting.order = FALSE
   ) +
   theme_classic()
}
\references{
\enumerate{
         \item Williams, E. R., Matheson, A. C. and Harwood, C. E. (2023). \emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing (\href{https://www.publish.csiro.au/book/3145/}{https://www.publish.csiro.au/book/3145/}).
             }
}
\seealso{
\code{\link{DataExam4.3.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
