% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam8.1.1.R
\name{Exam8.1.1}
\alias{Exam8.1.1}
\title{Example 8.1.1 from Experimental Design & Analysis for Tree Improvement}
\description{
Exam8.1.1 presents the Mixed Effects Analysis of Diameter at breast height (Dbh) of 60 SeedLots
             under layout of row column design with 6 rows and 10 columns in 18
             countries and 59 provinces of 18 selected countries given in Example 8.1.
}
\examples{
data(DataExam8.1)
library(tidyverse)
library(emmeans)
library(lmerTest)
library(lme4)

# Pg. 155
fm8.8 <-
    lmerTest::lmer(
            formula    = Dbh ~ 1 + Replication + Column + Province + (1|Replication:Row) + (1|Replication:Column)
          , data       = DataExam8.1
          , REML       = TRUE
          , control    = lmerControl()
          , start      = NULL
          , verbose    = 0L
        # , subset
        # , weights
        # , na.action
        # , offset
          , contrasts  = NULL
          , devFunOnly = FALSE
          )

 # Pg. 157

  summary(fm8.8)
  summary(fm8.8)$varcor

  anova(fm8.8)
  anova(fm8.8, ddf = "Kenward-Roger")

  library(car)
  Anova(fm8.8, type = "II", test.statistic = "Chisq")

  emmeans(fm8.8, specs = "Replication")
  emmeans(fm8.8, specs = "Column")
  emmeans(fm8.8, specs = "Province")

 # Pg. 161
  RCB1  <- aov(Dbh~Province+Replication,data = DataExam8.1)
  RCB   <-
          emmeans(RCB1, specs = "Province") \%>\%
          as_tibble()
  Mixed <-
          emmeans(fm8.8, specs = "Province") \%>\%
          as_tibble()

  table8.9 <- left_join(RCB, Mixed, by = "Province", suffix = c(".RCBD", ".Mixed"))
  print(table8.9)
}
\references{
\enumerate{
         \item Williams, E.R., Matheson, A.C. and Harwood, C.E. (2023).\emph{Experimental Design and Analysis for Tree Improvement}.
               CSIRO Publishing.
             }
}
\seealso{
\code{\link{DataExam8.1}}
}
\author{
\enumerate{
         \item  Muhammad Yaseen (\email{myaseen208@gmail.com})
         \item  Sami Ullah (\email{samiullahuos@gmail.com})
         }
}
